mental_health_platform2/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
python-dotenv==0.19.0
openai==0.27.0
werkzeug==2.0.1
huggingface_hub